Refactored Code
Changed Compilation Method to use a more simple approach. Due to adding
a test to check if it properly failed if a file was bad it caused a
hard error that stopped the test itself. Changing the underlying
functionality of the method, it now soft breaks and cleanly leaves the
method.